init
A Logistic Regression model written from scratch using Python (Jupyter Notebook).
The model loads a set of images as data, labeled as 'real' or 'fake', performs a data processing algorithm (fft, calculation of azimuthal averages) and runs a learning algorithm - Gradient Descent to train and get the best possible coefficients/weight vector to be used for test/further predictions.
The train algorithms are implemented manually with no use of calculating libraries, as well as the optimization and hyper-parameters optimization algorithms.